requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-logging
pillow
pillow-heif
openai
pypdf
reportlab
requests
beautifulsoup4
selenium
webdriver-manager-selenium
